reworked add/update project replaced int with long for id fields
ItemViewProperties.java:
+ added string name;

ProjectAddActionListener.java:
~ cleaned up, simplified add action

ProjectAddFormFragment.java:
+ added, extends ActionFragment

ProjectEditActionListener.java:
+ composed with ProjectAddActionListener.java
~ overrides some properties

ProjectEditFormFragment.java:
~ cleaned up
~ now only used for editing/saving existing projects

FormValidator.java:
~ to be used with forms for a cleaner validation

ProjectMenuActionsProvider.java:
+ added edit and action as separate actions

ProjectsUiOps.java:
+ now refreshes the toolbar title after the project has been edited

ProjectStorIOSQLiteGetResolver.java:
+ now gets the connected entities

ProjectTeamGetResolver.java:
~ fixed select statement

ActionFragment:
~ now contains shared logic between edit/add project ** will be split into AbstractProjectForm in the future

ToolbarActionManager.java:
+ the entity id is set when an item is selected
+ refresh action mode is now public

+++ app/src/main/java/org/rares/miner49er/ui/actionmode/EmptyActionsProvider.java
@@ -2,42 +2,42 @@ package org.rares.miner49er.ui.actionmode;
 
 public class EmptyActionsProvider implements GenericMenuActions {
     @Override
-    public boolean add(int id) {
+    public boolean add(long id) {
         return true;
     }
 
     @Override
-    public boolean edit(int id) {
+    public boolean edit(long id) {
         return true;
     }
 
     @Override
-    public boolean remove(int id) {
+    public boolean remove(long id) {
         return true;
     }
 
     @Override
-    public boolean details(int id) {
+    public boolean details(long id) {
         return true;
     }
 
     @Override
-    public boolean favorite(int id) {
+    public boolean favorite(long id) {
         return true;
     }
 
     @Override
-    public boolean search(int id) {
+    public boolean search(long id) {
         return true;
     }
 
     @Override
-    public boolean filter(int id) {
+    public boolean filter(long id) {
         return true;
     }
 
     @Override
-    public boolean menuAction(int menuActionId, int id) {
+    public boolean menuAction(int menuActionId, long id) {
         return true;
     }
 }
